+++ src/app/+state/auth/auth.selectors.ts
@@ -1,5 +1,5 @@
 import { createSelector } from '@ngrx/store';
-import { AppState } from '.';
+import { AppState } from '..';
 
 export const selectAuthToken = (state: AppState) => state.auth.token;
 export const selectAuthUser = (state: AppState) => state.auth.user;
