+++ src/app/+state/courses/courses.selectors.ts
@@ -1,4 +1,4 @@
-import { AppState } from '.';
+import { AppState } from '..';
 
 export const selectCourses = (state: AppState) => state.courses.courses;
 export const selectEditedCourse = (state: AppState) => state.courses.course;
