+++ src/app/auth/login-page/login-page.component.ts
@@ -1,20 +1,19 @@
 import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { login } from 'src/app/+state/auth.actions';
+import { login } from '../../+state/auth/auth.actions';
 
 @Component({
   selector: 'cl-login-page',
   templateUrl: './login-page.component.html',
-  styleUrls: ['./login-page.component.scss']
+  styleUrls: ['./login-page.component.scss'],
 })
 export class LoginPageComponent implements OnInit {
   email: string = '';
   password: string = '';
 
-  constructor(private store: Store) { }
+  constructor(private store: Store) {}
 
-  ngOnInit(): void {
-  }
+  ngOnInit(): void {}
 
   login() {
     this.store.dispatch(login({ email: this.email, password: this.password }));
